chore: remove noir_js_backend_barretenberg
