+++ barretenberg/ts/package.json
@@ -55,6 +55,7 @@
     "comlink": "^4.4.1",
     "commander": "^10.0.1",
     "debug": "^4.3.4",
+    "fflate": "^0.8.0",
     "tslib": "^2.4.0"
   },
   "devDependencies": {
