+++ barretenberg/ts/src/index.ts
@@ -8,4 +8,4 @@ export {
   UltraHonkBackend,
 } from './barretenberg/index.js';
 export { RawBuffer, Fr } from './types/index.js';
-export { splitHonkProof, reconstructHonkProof } from './proof/index.js';
+export { splitHonkProof, reconstructHonkProof, ProofData } from './proof/index.js';
