--- noir/noir-repo/.github/scripts/backend-barretenberg-build.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-set -eu
-
-yarn workspace @noir-lang/backend_barretenberg build
