+++ noir/noir-repo/.github/workflows/test-js-packages.yml
@@ -73,7 +73,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: noirc_abi_wasm
-          path: | 
+          path: |
             ./tooling/noirc_abi_wasm/nodejs
             ./tooling/noirc_abi_wasm/web
           retention-days: 10
@@ -263,9 +263,6 @@ jobs:
       - name: Build noir_js_types
         run: yarn workspace @noir-lang/types build
 
-      - name: Build barretenberg wrapper
-        run: yarn workspace @noir-lang/backend_barretenberg build
-
       - name: Run noir_js tests
         run: |
           yarn workspace @noir-lang/noir_js build
@@ -416,7 +413,7 @@ jobs:
       - name: Setup `integration-tests`
         run: |
           # Note the lack of spaces between package names.
-          PACKAGES_TO_BUILD="@noir-lang/types,@noir-lang/backend_barretenberg,@noir-lang/noir_js"
+          PACKAGES_TO_BUILD="@noir-lang/types,@noir-lang/noir_js"
           yarn workspaces foreach -vtp --from "{$PACKAGES_TO_BUILD}" run build
 
       - name: Run `integration-tests`
@@ -461,7 +458,7 @@ jobs:
       - name: Setup `integration-tests`
         run: |
           # Note the lack of spaces between package names.
-          PACKAGES_TO_BUILD="@noir-lang/types,@noir-lang/backend_barretenberg,@noir-lang/noir_js"
+          PACKAGES_TO_BUILD="@noir-lang/types,@noir-lang/noir_js"
           yarn workspaces foreach -vtp --from "{$PACKAGES_TO_BUILD}" run build
 
       - name: Run `integration-tests`
@@ -565,7 +562,7 @@ jobs:
     runs-on: ubuntu-latest
     # We want this job to always run (even if the dependant jobs fail) as we want this job to fail rather than skipping.
     if: ${{ always() }}
-    needs: 
+    needs:
       - test-acvm_js-node
       - test-acvm_js-browser
       - test-noirc-abi
