+++ noir/noir-repo/compiler/integration-tests/test/browser/recursion.test.ts
@@ -5,7 +5,7 @@ import { Logger } from 'tslog';
 import { acvm, abi, Noir } from '@noir-lang/noir_js';
 
 import * as TOML from 'smol-toml';
-import { BarretenbergBackend } from '@noir-lang/backend_barretenberg';
+import { UltraPlonkBackend } from '@aztec/bb.js';
 import { getFile } from './utils.js';
 import { Field, InputMap } from '@noir-lang/noirc_abi';
 import { createFileManager, compile } from '@noir-lang/noir_wasm';
@@ -73,7 +73,7 @@ describe('It compiles noir program code, receiving circuit bytes and abi object.
 
     const recursion_program = await getCircuit(`${base_relative_path}/${circuit_recursion}`);
 
-    const recursion_backend = new BarretenbergBackend(recursion_program);
+    const recursion_backend = new UltraPlonkBackend(recursion_program.bytecode);
 
     const { witness: recursion_witnessUint8Array } = await new Noir(recursion_program).execute(recursion_inputs);
