+++ noir/noir-repo/package.json
@@ -8,7 +8,6 @@
     "tooling/noirc_abi_wasm",
     "tooling/noir_js",
     "tooling/noir_codegen",
-    "tooling/noir_js_backend_barretenberg",
     "acvm-repo/acvm_js",
     "docs"
   ],
