--- noir/noir-repo/tooling/noir_js_backend_barretenberg/.eslintignore
@@ -1,1 +1,0 @@
-node_modules
