--- noir/noir-repo/tooling/noir_js_backend_barretenberg/.eslintrc.cjs
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: ["../../.eslintrc.js"],
-};
