--- noir/noir-repo/tooling/noir_js_backend_barretenberg/.gitignore
@@ -1,2 +1,0 @@
-crs
-lib
