--- noir/noir-repo/tooling/noir_js_backend_barretenberg/.mocharc.json
@@ -1,11 +1,0 @@
-{
-  "require": "ts-node/register",
-  "loader": "ts-node/esm",
-  "extensions": [
-    "ts",
-    "cjs"
-  ],
-  "spec": [
-    "test/**/*.test.ts*"
-  ]
-}
