--- noir/noir-repo/tooling/noir_js_backend_barretenberg/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "@noir-lang/backend_barretenberg",
-  "contributors": [
-    "The Noir Team <team@noir-lang.org>"
-  ],
-  "version": "0.35.0",
-  "packageManager": "yarn@3.5.1",
-  "license": "(MIT OR Apache-2.0)",
-  "type": "module",
-  "homepage": "https://noir-lang.org/",
-  "repository": {
-    "url": "https://github.com/noir-lang/noir.git",
-    "directory": "tooling/noir_js_backend_barretenberg",
-    "type": "git"
-  },
-  "bugs": {
-    "url": "https://github.com/noir-lang/noir/issues"
-  },
-  "source": "src/index.ts",
-  "main": "lib/cjs/index.js",
-  "module": "lib/esm/index.js",
-  "exports": {
-    "require": "./lib/cjs/index.js",
-    "types": "./lib/esm/index.d.ts",
-    "default": "./lib/esm/index.js"
-  },
-  "types": "lib/esm/index.d.ts",
-  "files": [
-    "lib",
-    "package.json"
-  ],
-  "scripts": {
-    "dev": "tsc --watch",
-    "generate:package": "bash ./fixup.sh",
-    "build": "yarn clean && tsc && tsc -p ./tsconfig.cjs.json && yarn generate:package",
-    "clean": "rm -rf ./lib",
-    "prettier": "prettier 'src/**/*.ts'",
-    "prettier:fix": "prettier --write 'src/**/*.ts' 'test/**/*.ts'",
-    "nightly:version": "jq --arg new_version \"-$(git rev-parse --short HEAD)$1\" '.version = .version + $new_version' package.json > package-tmp.json && mv package-tmp.json package.json",
-    "publish": "echo 📡 publishing `$npm_package_name` && yarn npm publish",
-    "lint": "NODE_NO_WARNINGS=1 eslint . --ext .ts --ignore-path ./.eslintignore  --max-warnings 0"
-  },
-  "dependencies": {
-    "@aztec/bb.js": "portal:../../../../barretenberg/ts",
-    "@noir-lang/types": "workspace:*",
-    "fflate": "^0.8.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20.6.2",
-    "@types/prettier": "^3",
-    "eslint": "^8.57.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "prettier": "3.2.5",
-    "ts-node": "^10.9.1",
-    "typescript": "5.4.2"
-  }
-}
