--- noir/noir-repo/tooling/noir_js_backend_barretenberg/src/index.ts
@@ -1,6 +1,0 @@
-export { BarretenbergBackend, UltraHonkBackend } from './backend.js';
-export { BarretenbergVerifier, UltraHonkVerifier } from './verifier.js';
-
-// typedoc exports
-export { Backend, CompiledCircuit, ProofData } from '@noir-lang/types';
-export { BackendOptions } from '@aztec/bb.js';
