--- noir/noir-repo/tooling/noir_js_backend_barretenberg/tsconfig.cjs.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "CommonJS",
-    "moduleResolution": "Node",
-    "outDir": "./lib/cjs"
-  },
-}
