--- noir/noir-repo/tooling/noir_js_backend_barretenberg/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext",
-    "declaration": true,
-    "emitDeclarationOnly": false,
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "outDir": "./lib/esm",
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "strict": true,
-    "noImplicitAny": false,
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules"],
-  "references": [
-    {
-      "path": "../noir_js_types"
-    }
-  ]
-}
